new system working, including split and merge pages are leaked on some octree_writes, though (chunk unloading is likely to do that)

+++ voxel_game/util/virtual_allocator.hpp
@@ -65,7 +65,7 @@ public:
 	inline uintptr_t indexof (T* ptr) const {
 		return ptr - (T*)baseptr;
 	}
-	inline T* operator[] (uint32_t index) const {
+	inline T* operator[] (uintptr_t index) const {
 		return (T*)baseptr + index;
 	}
